Moves UserRepo to ShoeboxModule
