+++ kifi-backend/modules/shoebox/test/com/keepit/model/SocialConnectionTest.scala
@@ -7,12 +7,12 @@ import java.io.File
 import org.specs2.mutable._
 import play.api.libs.json._
 import com.keepit.common.net.FakeHttpClientModule
-import com.keepit.common.store.FakeStoreModule
+import com.keepit.common.store.FakeShoeboxStoreModule
 import com.google.inject.Injector
 
 class SocialConnectionTest extends Specification with ShoeboxTestInjector {
 
-  val socialConnectionTestModules = Seq(FakeHttpClientModule(), FakeStoreModule())
+  val socialConnectionTestModules = Seq(FakeHttpClientModule(), FakeShoeboxStoreModule())
 
   private def extractFacebookFriendInfo(json: JsValue)(implicit injector: Injector): Seq[(SocialUserInfo, JsValue)] = {
     inject[FacebookSocialGraph].extractFriends(json)
