+++ kifi-backend/modules/shoebox/test/com/keepit/search/ArticleSearchResultTest.scala
@@ -7,12 +7,13 @@ import com.keepit.test._
 import org.specs2.mutable._
 import play.api.test.Helpers._
 import com.keepit.serializer.ArticleSearchResultSerializer
+import com.google.inject.Injector
 
-class ArticleSearchResultTest extends Specification with ApplicationInjector with DeprecatedDbRepos {
+class ArticleSearchResultTest extends Specification with ShoeboxTestInjector {
 
   "ArticleSearchResult" should {
     "be serialized" in {
-      running(new DeprecatedEmptyApplication()) {
+      withDb() { implicit injector: Injector =>
          val res = ArticleSearchResult(
               last = Some(ExternalId[ArticleSearchResultRef]()),
               query = "scala query",
@@ -34,7 +35,7 @@ class ArticleSearchResultTest extends Specification with ApplicationInjector wit
     }
 
     "persisting to db" in {
-      running(new DeprecatedEmptyApplication()) {
+      withDb() { implicit injector: Injector =>
         val repo = inject[ArticleSearchResultRefRepo]
          val user = db.readWrite { implicit s =>
            userRepo.save(User(firstName = "Shachaf", lastName = "Smith"))
