+++ kifi-backend/modules/shoebox/test/com/keepit/social/SocialUserImportFriendsTest.scala
@@ -9,16 +9,19 @@ import org.specs2.mutable._
 import com.keepit.common.db.Id
 import com.keepit.inject._
 import com.keepit.model.SocialUserInfo
-import com.keepit.test.DeprecatedEmptyApplication
+import com.keepit.test.{ShoeboxTestInjector, TestInjector, DeprecatedEmptyApplication}
 
 import play.api.libs.json.Json
 import play.api.test.Helpers._
+import com.google.inject.Injector
+import com.keepit.common.net.FakeHttpClientModule
+import com.keepit.common.store.FakeShoeboxStoreModule
 
-class SocialUserImportFriendsTest extends Specification with ApplicationInjector {
+class SocialUserImportFriendsTest extends Specification with ShoeboxTestInjector {
 
   "SocialUserImportFriends" should {
     "import friends" in {
-      running(new DeprecatedEmptyApplication().withFakeHttpClient()) {
+      withDb(FakeHttpClientModule(), FakeShoeboxStoreModule()) { implicit injector =>
         val graphs = List(
             ("facebook_graph_andrew_min.json", 7),
             ("facebook_graph_eishay_super_min.json", 5),
@@ -31,7 +34,7 @@ class SocialUserImportFriendsTest extends Specification with ApplicationInjector
     }
   }
 
-  def testFacebookGraph(jsonFilename: String, numOfFriends: Int) = {
+  def testFacebookGraph(jsonFilename: String, numOfFriends: Int)(implicit injector: Injector) = {
     val json = Json.parse(io.Source.fromFile(new File("modules/shoebox/test/com/keepit/common/social/data/%s".format(jsonFilename))).mkString)
     val extractedFriends = inject[FacebookSocialGraph].extractFriends(json)
     val rawFriends = inject[SocialUserImportFriends].importFriends(extractedFriends)
