+++ kifi-backend/modules/shoebox/test/com/keepit/social/UserConnectionCreatorTest.scala
@@ -7,17 +7,19 @@ import org.specs2.mutable._
 import com.keepit.common.db.slick.Database
 import com.keepit.inject.ApplicationInjector
 import com.keepit.model._
-import com.keepit.test.DeprecatedEmptyApplication
+import com.keepit.test.{ShoeboxTestInjector, DeprecatedEmptyApplication}
 
 import play.api.libs.json.Json
 import play.api.test.Helpers._
+import com.keepit.common.net.FakeHttpClientModule
+import com.keepit.common.store.FakeShoeboxStoreModule
 
-class UserConnectionCreatorTest extends Specification with ApplicationInjector {
+class UserConnectionCreatorTest extends Specification with ShoeboxTestInjector {
 
 
   "UserConnectionCreator" should {
     "create connections between friends" in {
-      running(new DeprecatedEmptyApplication().withFakeHttpClient()) {
+      withDb(FakeHttpClientModule(), FakeShoeboxStoreModule()) { implicit injector =>
 
         /*
          * grab json
@@ -66,7 +68,7 @@ class UserConnectionCreatorTest extends Specification with ApplicationInjector {
     }
 
     "disable non existing connections" in {
-      running(new DeprecatedEmptyApplication().withFakeHttpClient()) {
+      withDb(FakeHttpClientModule(), FakeShoeboxStoreModule()) { implicit injector =>
 
         val json1 = Json.parse(io.Source.fromFile(new File("modules/shoebox/test/com/keepit/common/social/data/facebook_graph_eishay_min.json")).mkString)
